fix(kno-2526): always checkpoint to most recent message (#7)
kcl_ex uses Broadway.CallerAcknowledger to acknowledge messages. This sends a set of success messages and a set of failed messages to the producer. The producer is then supposed to sort out what to do about this.

Long-term, this is not the right acknowledger for us to be using, since we just need the highwater mark for all messages in the batch. Broadway's Kafka adapter has a custom acknowledger for this reason.

Short-term, we are patching kcl_ex to exhibit the same behavior as the Kafka producer: Ack all messages and advance to the next set of messages in the shard, regardless of error outcomes. It is up to the consumer to handle_failed according to their appropriate use case.